Implemented new main menu, associated options menu and app icon.
This main menu displays a list of forms that can be used to start new
records.  If forms are not present the options menu is automatically
opened and a little hint is displayed telling the user how to add new
forms.  A short-lived hint about what will happen if the user clicks
on a form in the list is displayed if there are forms present.

The TFMainMenuActivity class may also be used as an intent to display
a list of installed forms to pick from.