Continued to work on form builder (FB), reading and writing XForm files.
FORM BUILDER FIELD LIST
========================================================================

This codebase now calls FormBuilderFieldEditor for a result.  As is
mentioned below, further post-save initialization and alteration of
the form objects is made here before committing them to the various
states.

Logic in gotoActiveField was split into gotoActiveField() and
gotoParentField() since the core purposes of this logic is different.

Fixed bug having to do with orientation changes & navigation (repeated
groups were not being respected).

FORM BUILDER FIELD EDITOR
========================================================================

Simplified the code base for this class.  It is responsible for the
creation of new fields & editing of existing ones.  The display should
be able to be rotated without any problems.

Logic is separated into onCreate() for initialization, per-field
`load' methods for further type-specific initialization, saveChanges()
and per-field `save' methods.  Nothing about the Field objects loaded
or created is committed until the saveChanges() routine is run.

Pressing the "back" key will cause a quit dialog to appear, identical
in function to the quit dialog seen on the form entry activity.

It is important to note that not all logic will be run to commit
changes as of the completion of saveChanges() and per-field `save'
methods: some field saving logic also exists in FormBuilderFieldList
in reaction to the outcome of the FormBuilderFieldEditor activity.

OTHER CHANGES
========================================================================

* FB field editor layout changes: hint input & related fields may now
  be disabled as a group; added group type selection inputs.

* Unified Instance, Bind and Field terminology.  ref, nodeset and
  xpath are now all called XPath.

* Added hack to FormReader so that newly created forms that fail the
  first save (and thus do not have an instance root) do not crash the
  application upon later FB load attempts; also added code to generate
  a usable instance root name if one cannot be derived from the form
  name.

* Fixed bug three bugs in FormWriter: duplicate itext tag removed; use
  of "nodeset" attribute for repeat tags is now determined on the
  basis of the field getType() output rather than getAttributes()
  which is not necessarily updated; temporarily disabled xmlns
  attribute on instance root tag (this was more difficult to implement
  than it first seemed).

* Changes to Field class:

  * Turned member variables attributes & children to private.  Updated
    all callers.

  * Switched to getRepeat() instead of the more cryptic
    getChildren().get(0) to access the repeat element of a repeated
    group.

  * Added static methods isRepeatedGroup() and makeFieldName() for
    convenience.

  * Removed `repeated' flag and added empty, newField and saved.  All
    three switches are using during initialization, saving and
    post-initialization of fields.