Implement Item30 (persistent sessions across restarts).
Reorganize the way that MainApplicationState deals with services.
This is a continued attempt to ensure that services start and stop
when they should.  This implementation is not complete: certain
services such as CouchDbService should not be stopped when the
application closes.

InformOnlineService became the preferred way to check for
connectivity, etc.  InformOnlineState may be used to check some things
but it should not be used to check connectivity state.  The difference
between the two will become more clear as time goes on
(InformOnlineState is becoming responsible for registration as stored
in preferences only).

Replace main browser window title app name (Group Inform) with button
that displays online/offline status and will (eventually) trigger a
connect/disconnect if pressed.

Properly encapsulate application-wide CouchDbService object accessor
in getDbService() and setDbService(), respectively.  Change all
callers.

Rename application-wide InformOnlineState object accessor to
getInformOnlineState and setInformOnlineState, respectively.  Change
all callers.