Fix support for retrying the connection to Inform Online for use cases where the device had not been specifically put into offline mode (post application init).
Complete rudimentary support for use cases where the user has
requested the device be offline.  Have yet to add a dialog that
reminds the user that the device is offline.

InitializeApplicationTask will never spend longer than 30 seconds
attempting to initialize the app.

Fix bug in InitializeApplicationTask where connections might not be
established as soon as expected due to InformOnlineService not having
been fully restarted and the usual 10 minutes between checks.

Fix bug where goOnline might accidentally try to connect while a
connection was in-progress.

Implement workaround in reinitializeService() for services not
stopping/being destroyed as expected.  This is also used when retrying
failed connections on behalf of the user.

Remove unneeded stopService() calls in cleanup().  These are only
needed if explicit startService() calls are used (they are not).

Other miscellaneous but not quite as important changes included.