Implement device reset functionality for account owners.
Refactored "self device reset" to use an AsyncTask.

Users that try to remove their default database will receive a message
telling them that removal of this special database is not permitted.

Alter device list adapter so that device registrations that are not
active are shown with a "blocked" icon.  This may need to be explained
better to the user.

Tidy up device layout, form layout and strings.