Implement CouchDB as per Item39.  Implemention is not bug free but at least we can get this major item off of our list.
Resolve Item25 (reset device functionality for Couch data).

Rename MainBrowserActivity to LauncherActivity to help better describe
the role this class plays.  Considering spliting the UI and
init/install functions into separate classes.

Renamed CouchDBService to DatabaseService to differentate it from
CouchService which is responsible for the Android CouchDB process.
DatabaseService continues to be responsible for connecting &
communciating with databases (but has nothing to do with the database
processes).

Upgrade to Ektorp 1.0.2 release.