Merge upstream changeset
changeset:   584:b4d13691d546
user:        carlhartung
date:        Sat Jun 11 15:12:44 2011 -0700
description:
Added ability to save-as.  that is, change the displayed name of a saved instance

Nothing is done with the instance name attribute at this point (it
isn't even set unless the user provides a name that is different than
the form definition name provided as a default).  A future change may
opt to display it in the form browser if it is present.