Merge upstream changeset as-is
changeset:   586:d6465b46cf19
user:        carlhartung
date:        Mon Jun 13 15:52:13 2011 -0700
description:
Fixes the bug where a swipe that brings up a constraint triggers a
'long-press' on a view.  The problem was that the 'up' action is
handled by the activity, so the view never sees it and pops the
long-press menu.  We fixed it by calling cancelLongPress() when we
notice the user scrolling.

This checkin also removes our home-grown gesture detector and uses
android's built in, instead.