Item113: Placeholders are not removed (also rename FormDefinition.Status.temporary to placeholder)
This update requires that all databases be updated to
FormDefinitionRepoR2 and FormInstanceRepoR2.

Client-based removal of placeholders occurs when the browser activity
is refreshed and will occur at most once every five minutes per
database.  The logic is as follows:

- remove all "old-style" placeholders (missing createdBy or dateCreated)
- remove all placeholders created by me
- remove all placeholders created by others that are stale (older than one day)

See wiki Projects.GroupComplete.RefactoringToDoList (rev 2) where it
states that we would like to remove the need for placeholders
alltogether since they are really more of a pain than anything.